+++ app/controllers/catalog_controller.rb
@@ -74,13 +74,16 @@ class CatalogController < ApplicationController
 
     config.view.list.thumbnail_field = :thumbnail_square_url_ssm
     config.view.list.partials = [:exhibits_document_header, :index]
+    config.view.gallery.title_only_by_default = true
     config.view.gallery.partials = [:index_header, :index]
     config.view.gallery.default_bibliography_thumbnail = 'default-square-thumbnail-book-large.png'
     config.view.gallery.default_canvas_thumbnail = 'default-square-thumbnail-annotation-large.png'
     config.view.heatmaps.partials = []
     config.view.heatmaps.color_ramp = ['#ffffcc', '#a1dab4', '#41b6c4', '#2c7fb8', '#253494']
+    config.view.masonry.title_only_by_default = true
     config.view.masonry.partials = [:index]
     config.view.masonry.default_bibliography_thumbnail = 'default-square-thumbnail-book-large.png'
+    config.view.slideshow.title_only_by_default = true
     config.view.slideshow.partials = [:index]
     config.view.embed.partials = [:viewer]
     config.view.embed.if = false